python/requirements.txt
pillow
tabulate
numpy
torch
transformers
scikit-learn
pandas
opencv-python
imagehash